+++ kernel-rs/src/fs/inode.rs
@@ -1,5 +1,5 @@
 use super::{
-    balloc, bfree, Dirent, FileName, BSIZE, DIRENT_SIZE, IPB, MAXFILE, NDIRECT, NINDIRECT,
+    Dirent, FileName, BSIZE, DIRENT_SIZE, BPB, MAXFILE, NDIRECT, NINDIRECT,
 };
 
 use crate::{
@@ -36,6 +36,9 @@ pub struct Inode {
 
     pub inner: Sleeplock<InodeInner>,
 }
+
+/// Inodes per block.
+pub const IPB: usize = BSIZE.wrapping_div(mem::size_of::<Dinode>());
 
 #[derive(Clone)]
 pub struct IcacheTag {}
@@ -330,7 +333,7 @@ impl InodeGuard<'_> {
 // which should follow C(=machine) representation
 // https://github.com/kaist-cp/rv6/issues/52
 #[repr(C)]
-pub struct Dinode {
+struct Dinode {
     /// File type
     typ: i16,
 
@@ -528,4 +531,48 @@ impl Inode {
             ),
         }
     }
+}
+
+/// Zero a block.
+unsafe fn bzero(dev: i32, bno: i32) {
+    let mut bp = Disk::read(dev as u32, bno as u32);
+    ptr::write_bytes(bp.deref_mut_inner().data.as_mut_ptr(), 0, BSIZE);
+    kernel().fs().log_write(bp);
+}
+
+/// Blocks.
+/// Allocate a zeroed disk block.
+unsafe fn balloc(dev: u32) -> u32 {
+    let mut bi: u32 = 0;
+    for b in num_iter::range_step(0, kernel().fs().superblock.size, BPB) {
+        let mut bp = Disk::read(dev, kernel().fs().superblock.bblock(b));
+        while bi < BPB && (b + bi) < kernel().fs().superblock.size {
+            let m = 1 << (bi % 8);
+            if bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & m == 0 {
+                // Is block free?
+                bp.deref_mut_inner().data[(bi / 8) as usize] =
+                    (bp.deref_mut_inner().data[(bi / 8) as usize] as i32 | m) as u8; // Mark block in use.
+                kernel().fs().log_write(bp);
+                bzero(dev as i32, (b + bi) as i32);
+                return b + bi;
+            }
+            bi += 1;
+        }
+    }
+    panic!("balloc: out of blocks");
+}
+
+/// Free a disk block.
+unsafe fn bfree(dev: i32, b: u32) {
+    let mut bp = Disk::read(dev as u32, kernel().fs().superblock.bblock(b));
+    let bi: i32 = b.wrapping_rem(BPB) as i32;
+    let m: i32 = (1) << (bi % 8);
+    assert_ne!(
+        bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & m,
+        0,
+        "freeing free block"
+    );
+    bp.deref_mut_inner().data[(bi / 8) as usize] =
+        (bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & !m) as u8;
+    kernel().fs().log_write(bp);
 }
