+++ kernel-rs/src/fs/mod.rs
@@ -12,7 +12,6 @@
 /// On-disk file system format used for both kernel and user programs are also included here.
 use crate::{
     bio::Buf,
-    kernel::kernel,
     log::Log,
     sleepablelock::Sleepablelock,
     stat::T_DIR,
@@ -24,7 +23,7 @@ use core::{mem, ptr};
 mod path;
 pub use path::{FileName, Path};
 mod inode;
-pub use inode::{Dinode, Inode, InodeGuard, InodeInner, RcInode};
+pub use inode::{Inode, InodeGuard, InodeInner, RcInode, IPB};
 
 /// Disk layout:
 /// [ boot block | super block | log | inode blocks |
@@ -115,9 +114,6 @@ const NDIRECT: usize = 12;
 const NINDIRECT: usize = BSIZE.wrapping_div(mem::size_of::<u32>());
 const MAXFILE: usize = NDIRECT.wrapping_add(NINDIRECT);
 
-/// Inodes per block.
-const IPB: usize = BSIZE.wrapping_div(mem::size_of::<Dinode>());
-
 impl Superblock {
     /// Block containing inode i
     const fn iblock(self, i: u32) -> u32 {
@@ -197,48 +193,4 @@ impl FileSystem {
     pub unsafe fn log_write(&self, b: Buf) {
         self.log.lock().log_write(b);
     }
-}
-
-/// Zero a block.
-unsafe fn bzero(dev: i32, bno: i32) {
-    let mut bp = Disk::read(dev as u32, bno as u32);
-    ptr::write_bytes(bp.deref_mut_inner().data.as_mut_ptr(), 0, BSIZE);
-    kernel().fs().log_write(bp);
-}
-
-/// Blocks.
-/// Allocate a zeroed disk block.
-unsafe fn balloc(dev: u32) -> u32 {
-    let mut bi: u32 = 0;
-    for b in num_iter::range_step(0, kernel().fs().superblock.size, BPB) {
-        let mut bp = Disk::read(dev, kernel().fs().superblock.bblock(b));
-        while bi < BPB && (b + bi) < kernel().fs().superblock.size {
-            let m = 1 << (bi % 8);
-            if bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & m == 0 {
-                // Is block free?
-                bp.deref_mut_inner().data[(bi / 8) as usize] =
-                    (bp.deref_mut_inner().data[(bi / 8) as usize] as i32 | m) as u8; // Mark block in use.
-                kernel().fs().log_write(bp);
-                bzero(dev as i32, (b + bi) as i32);
-                return b + bi;
-            }
-            bi += 1;
-        }
-    }
-    panic!("balloc: out of blocks");
-}
-
-/// Free a disk block.
-unsafe fn bfree(dev: i32, b: u32) {
-    let mut bp = Disk::read(dev as u32, kernel().fs().superblock.bblock(b));
-    let bi: i32 = b.wrapping_rem(BPB) as i32;
-    let m: i32 = (1) << (bi % 8);
-    assert_ne!(
-        bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & m,
-        0,
-        "freeing free block"
-    );
-    bp.deref_mut_inner().data[(bi / 8) as usize] =
-        (bp.deref_mut_inner().data[(bi / 8) as usize] as i32 & !m) as u8;
-    kernel().fs().log_write(bp);
 }
